Improve helper script for making release. [skip ci]
* Merge source tarball generation script.
* Generate Python source wheel.
* Generate hashes and release note.
